+++ project.py
@@ -1,5 +1,5 @@
-from pickle import TRUE
 import pygame
+from utils.sorts import *
 import math as mt
 import random as rd
 
@@ -7,10 +7,11 @@ pygame.init()
 
 
 class SortingVisualizer:
+    WHITE = 255, 255, 255
+    # BG_COLOR = WHITE
     BG_COLOR = 0, 0, 0
     # BG_COLOR = 247, 147, 30  # GOLD
     # BG_COLOR = 128, 128, 128  # GREY
-    WHITE = 255, 255, 255
     BLACK = 0, 0, 0
     SIDE_MARGIN = 100
     TOP_MARGIN = 100
@@ -23,7 +24,7 @@ class SortingVisualizer:
         (58, 190, 255),
     ]
     FONT = pygame.font.SysFont("verdana", 30)
-    MENU_FONT = pygame.font.SysFont("verdana", 20)
+    MENU_FONT = pygame.font.SysFont("verdana", 18)
 
     def __init__(self, width, height, lst):
         self.width = width
@@ -49,7 +50,7 @@ def display_handler(visualizer, sort_name, ascending):
     visualizer.window.blit(
         current_sort, (visualizer.width // 2 - current_sort.get_width() // 2, 5))
     menu = visualizer.MENU_FONT.render(
-        "X - Reset | Enter - Start | A - Ascending | D - Descending | B - Bubble Sort | I - Insertion Sort", 1, visualizer.WHITE)
+        "X - Reset | Enter - Start | A - Ascending | D - Descending | B - Bubble Sort | I - Insertion Sort | S - Selection Sort", 1, visualizer.WHITE)
     visualizer.window.blit(
         menu, (visualizer.width // 2 - menu.get_width() // 2, 45))
     display_array_handler(visualizer)
@@ -85,41 +86,6 @@ def display_array_handler(visualizer, swappers={}, refresh=False):
         pygame.display.update()
 
 
-def bubbleSort(visualizer, ascending=True):
-    lst = visualizer.lst
-    for i in range(len(lst) - 1):
-        for j in range(len(lst) - i - 1):
-            n1 = lst[j]
-            n2 = lst[j+1]
-            if (n1 > n2 and ascending) or (n1 < n2 and not ascending):
-                lst[j], lst[j+1] = lst[j+1], lst[j]
-                display_array_handler(visualizer, swappers={
-                                      j: visualizer.RED, j+1: visualizer.YELLOW}, refresh=True)
-                yield True
-    return lst
-
-
-def insertionSort(visualizer, ascending=True):
-    lst = visualizer.lst
-    for i in range(1, len(lst)):
-        temp = lst[i]
-        j = i - 1
-        while True:
-            ascending_sort = i > 0 and lst[i-1] > temp and ascending
-            descending_sort = i > 0 and lst[i-1] < temp and not ascending
-            if not ascending_sort and not descending_sort:
-                break
-
-            lst[i] = lst[i-1]
-            i -= 1
-            lst[i] = temp
-            display_array_handler(visualizer, swappers={
-                j: visualizer.RED, j+1: visualizer.YELLOW}, refresh=True)
-            yield True
-
-    return lst
-
-
 def main_loop():
     isRunning = True
     clock = pygame.time.Clock()
@@ -136,7 +102,7 @@ def main_loop():
     current_sort_name = "Bubble Sort"
     current_sort_func = None
     lst = create_array(curr_min, curr_max, curr_range)
-    visualizer = SortingVisualizer(1000, 600, lst)
+    visualizer = SortingVisualizer(1100, 600, lst)
 
     while isRunning:
         clock.tick(60)
@@ -160,6 +126,11 @@ def main_loop():
                 lst = create_array(curr_min, curr_max, curr_range)
                 visualizer.update_array(lst)
                 isSorting = False
+
+            # Rare usage case.
+            if event.key == pygame.K_ESCAPE:
+                isRunning = False
+
             elif event.key == pygame.K_a and not isSorting:
                 ascending = True
             elif event.key == pygame.K_d and not isSorting:
@@ -171,6 +142,9 @@ def main_loop():
             elif event.key == pygame.K_i and not isSorting:
                 current_sort = insertionSort
                 current_sort_name = "Insertion Sort"
+            elif event.key == pygame.K_s and not isSorting:
+                current_sort = selectionSort
+                current_sort_name = "Selection Sort"
             elif event.key == pygame.K_b and not isSorting:
                 current_sort = bubbleSort
                 current_sort_name = "Bubble Sort"
